Translation of kinds and types
We are finally getting more idiomatic. Far from perfect of course, but
getting the ball rolling with some dignity. The next thing will be
fixed point definitions, where we will need to keep track of previous
definitions to generate the right mu/nu formulas. To be noted is that
we may need to generate several files to fit into our framework.