Parser stub for interactive inductive definitions
The Define category of inductive definitions is refined with a simple
"printer parser," alas without indentation, that demonstrates the
feasibility of the approach (as opposed to the new "ship" tactic
developed by Kaustuv in parallel, which is essentially a "printer
lexer" that translates symbols in sequence: this works for variable
renumbering, but it cannot be extended to our non-trivially-shallow
translation without instrumenting it with symbol tables and basically
all the trappings of a full-fledged parser).

As running examples, I am using addition and the mutually recursive
definition of parity, both taken from the Abella tutorial.