Equality and documentation
With the aforementioned limitations (cf. previous commits), equality is
now supported. The code has been cleaned up and commented on the grounds
that it will be very important to know exactly why I wrote something a
certain way and the assumptions and limitations of functions for their
extension and reuse. Lacking extensive test coverage at the moment (to
put it mildly), this seems the least that must be done even in the
current prototypical state of affairs.