Handling of declared identifiers in clause heads
A bit of imperative programming (ick) to reuse the Abella tables and
decide which identifiers are not variables in a clause head.

Note that the code movement makes it difficult to isolate the changes.
It may be advisable to split such commits in two: one to move code
around harmlessly, the second to implement the new functions.