Full inductive definitions
Experimenting with a more modular, pipe-based style of coding that also
seems more idiomatic. This is still experimental, and more cleanup is
needed.

However, definitions appear to be correct. Note that only interactive
treatment is being considered. Eventually we will want to shift to
compiled files, or do both while reusing as much code as possible.