Unified compiled-interactive translation (I)
Note: This snapshot still breaks!

The code is in the process of being transformed into a unified
implementation that considers all declarations as compiled (taking
advantage of the fact that interactive Abella sessions always end up
compiling everything, we can avoid the interactive half altogether and
accummulate the checks done during compilating to ensure consistency).

Some of the term structures for compiled Abella must be re-evaluated,
as they differ rather deeply from the interactive ones.