Full translation and file structuring
New register-certify API used uniformly throughout the main Abella file.
Now each relevant command is memoized and translated in bulk at the end
of the process. Mainly, we need to keep this information for the signature,
so we extend its use uniformly, and thus are able to generate all the
stubs within Abella.