Simplified module interface
The register-certify paradigm supersedes the command-by-command
treatment without loss of generality. All commands can be uniformly
handled by register. Fully interactive translation (which has some
problems WITH THE CURRENT CHECKER, as we know) becomes more tedious,
but a trivial extension to the interface can remedy this. On the
other hand, a centralized repository of operations is handier to have.