Use new "simple FPC" for generated harnesses
The more generic "admin FPC" used heretofore is replaced with the
newer, simpler, "simple FPC." This affords easier programming and
uses the current structure of signatures-modules incorporated into
the checker. Furthermore, a "simple certificate" is provided as a
placeholder for further experimentation, so that results can be run
out of the box and many simple results will be proved directly
without need for further changes.

Down the line, the FPC that we use should probably be selectable,
but the "simple FPC" is certainly a good starting point from the
user point of view.

The next step is to compare this automatic integration with what
results from manual encoding: are there any relevant differences,
and if so which? Recall that the automatic encoding differs
somewhat from that handcrafted for conciseness, and this may
(should?) be the principal source of any divergences.