mutual is a typed Itab vs. a list
The mutual table is presently a map of identifiers to their types, as
opposed to the list of identifiers of yore. In simple cases, i.e.,
involving "foo", we know the type from its declaration in Test_helper,
but this is neither clean not extensible to the more general maps
defined modularly for use with the Search and Unfold blocks. For this
reason, they will always be read from the signature, with the benefit
that if something goes wrong we won't be able to find the objects there,
and die.