Extend syntax of Specification with namespace
An alternative to having two separate, parallel import mechanism is
extending standard system functionality to support separate, though
combinable, namespaces. The idea here is that instead of

  Specification "foo".

to load a lProlog .sig and .mod file pair exactly at the beginning
of an Abella session, this behavior is preserved, but, optionally,
the user can specify an identifier that will serve as namespace for
the contents of the loaded module:

  Specification foo := "bar".

Each namespace will be loadable exactly once, without overwrites,
and identifiers will be shared with the top namespace to avoid
confusion. This default namespace continues to work as before,
whereas named namespaces can be loaded at any time in the session.

Curly brackets continue to operate by default on the top namespace,
also as usual. To refer to one particular namespace, the brackets
are prefixed with the corresponding identifier:

  foo{...}

All this works on lProlog file pairs, as before. A further
refinement would allow several specification languages to coexist:

  Specification foo ["spec"] := "bar.ext".

where lProlog would remain the default, and there would be a set of
modules in charge of compilation from specification languages to
internal Abella terms.