Process specs based on namespace
First effective changes to manipulate multiple namespaces in parallel.
The semantics of aggregation is still missing, but the fundamental
workflow works: namely, the default specification is only accepted at
the beginning of the session; all others can be given at any time if
they have not been defined before. (Currently, everything is processed
and ends up in the single default namespace.)