Pass certificate to the kernel
We define a nondefinitive predicate, test_cert : cert -> bool -> o, for
the `certify` tactic to pass a certificate and a kernel-level object
formula, to be translated from the reasoning level in abella, for
certification. In addition to preparing both certificate and formula in
sequent form, the predicate is in charge of polarization, as defined by
the user.