Unification error during case analysis: dead code
In `try_left_unify_cpairs`, the exception `UnifyError` is captured and
interpreted as a failure in unification error during case analysis.
However, the handler will catch these errors in `unify` and add their
source, necessarily of type `NotLLambda`, to the list of conflict pairs.
The code appears to be unused. If there are special conditions that
exercise it, they are not specified, as illustrated by this commit,
which changes this handling without any effect on the test suite.